fix require issues
