+++ src/rippled/ledgers.js
@@ -1,4 +1,4 @@
-import { summarizeLedger } from './lib/utils'
+import summarizeLedger from './lib/summarizeLedger'
 import { getLedger as getRippledLedger } from './lib/rippled'
 import logger from './lib/logger'
